Validating nullable data

+++ app/Http/Controllers/MarcaController.php
@@ -7,6 +7,10 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,8 +18,8 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        $marcas = Marca::all();
-        return $marcas;
+        $marcas = $this->marca->all();
+        return response($marcas, 200);
     }
 
     /**
@@ -36,19 +40,23 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca = Marca::create($request->all());
+        $marca = $this->marca->create($request->all());
 
-        return $marca;
+        return response($marca, 201);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null) {
+            return response(['erro' => 'Recurso solicitado não existe'],404);
+        }
         return $marca;
     }
 
@@ -67,23 +75,37 @@ class MarcaController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null) {
+            return response([
+                'erro' => 'Impossivel realizar a atualização, o recurso ' . $id . ' não existe'
+            ], 404);
+        }
+
         $marca->update($request->all());
+
         return $marca;
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null) {
+            return response()->json(
+                ['erro' => 'Impossivel realizar a atualização, o recurso ' . $id . ' não existe'], 404
+            );
+        }
         $marca->delete();
         return ['msg' => 'Marca removida com sucesso!'];
     }
